requirements.txt
streamlit==1.32.0
requests
beautifulsoup4==4.12.3
openai>=1.14.0
Pillow>=10.2.0
python-dotenv
playwright
